Enhance agent configuration and local API server integration
- Added per-agent request timeouts for local models to improve performance and prevent long hangs during processing.
- Updated the `configure_hermes_agent` function to seed a tighter request timeout for the custom provider, ensuring faster failure on issues.
- Removed the local API server option from the navigation and adjusted related components to streamline the user interface.
- Enhanced localization for the launch page, including new descriptions and labels for the local API server section.
- Refactored the launch page to improve state management and user feedback during agent installation and configuration processes.

These changes aim to optimize the integration of coding agents with local models, providing a smoother user experience and clearer feedback during operations.
